code-examples/chapter-04-security/complete/pyproject.toml
[project]
name = "secure-agent"
version = "0.1.0"
description = "Production-ready AI agent with comprehensive security"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "structlog>=24.1.0",
    "bleach>=6.1.0",
    # Observability (from Chapter 3)
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    # Optional: Secrets management
    "boto3>=1.34.0",  # AWS Secrets Manager
    "hvac>=2.1.0",    # HashiCorp Vault
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]
